modify emp

+++ S9_Day1Api/Models/Employee.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Salary { get; set; }
+        public double Salary { get; set; }
         public Department? Department { get; set; }
     }
 }
